Move Metric definition to sql package
